+++ PlantEvolution/app/src/main/java/ploobs/plantevolution/Model/Model3D/Vertices.java
@@ -111,7 +111,8 @@ public class Vertices
             float pointX, float pointY, float pointZ,
             float textureU, float textureV,
             float normalX, float normalY, float normalZ,
-            float colorR, float colorG, float colorB, float colorA)
+          //  float colorR, float colorG, float colorB, float colorA)
+            short colorR, short colorG, short colorB, short colorA)
     {
         _points.add(pointX, pointY, pointZ);
